[ARM] For arm32ve physical_addr_t and physical_size_t types are 64-bit due to use of LPAE.
This patch changes physica_addr_t and physical_size_t for cortex-a15-ve support.
It also fixed related compiler warnings.

+++ arch/arm/cpu/arm32ve/include/arch_types.h
@@ -27,8 +27,8 @@
 typedef unsigned int irq_flags_t;
 typedef unsigned int virtual_addr_t;
 typedef unsigned int virtual_size_t;
-typedef unsigned int physical_addr_t;
-typedef unsigned int physical_size_t;
+typedef unsigned long long physical_addr_t;
+typedef unsigned long long physical_size_t;
 typedef unsigned int clock_freq_t;
 typedef unsigned long long jiffies_t;
